requirements.txt
selenium==4.15.0
requests==2.31.0
pytest==7.4.3
allure-pytest==2.13.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
pytest-xdist==3.5.0
pytest-html==4.1.1
pytest-timeout==2.2.0
Faker==20.1.0
flake8==6.1.0